requirements.txt
requests
lxml
torch
transformers
joblib
faiss-cpu
